+++ src/test/java/com/jwplayer/jwplatform/exception/TestJWPlatformClientExceptions.java
@@ -5,7 +5,7 @@ import static org.mockito.Matchers.anyString;
 import static org.mockito.Mockito.when;
 import static org.powermock.api.mockito.PowerMockito.mockStatic;
 
-import com.jwplayer.jwplatform.JWPlatformClient;
+import com.jwplayer.jwplatform.v1.JWPlatformClient;
 import com.mashape.unirest.http.HttpResponse;
 import com.mashape.unirest.http.JsonNode;
 import com.mashape.unirest.http.Unirest;
